Refactor object-versions and widgets utils
Move getWidgetConfig() from object-version-utils to widgets-utils.
Update widgets-utils to work with object-version and mega-object.
