+++ src/ggrc-client/js/plugins/utils/object-versions-utils.js
@@ -8,7 +8,6 @@ import * as businessModels from '../../models/business-models';
 /**
    * Util methods for work with Object versions.
    */
-let cachedObjects = {};
 let modelsIncludeVersions = [
   'Issue',
 ];
@@ -35,74 +34,26 @@ function parentHasObjectVersions(parentModelName) {
   return modelsIncludeVersions.indexOf(parentModelName) > -1;
 }
 
-function _getObjectVersionConfig(modelName, forceBuildFromOriginal) {
+function getObjectVersionConfig(modelName) {
   let originalModelName;
   let objectVersion = {};
-  if (!forceBuildFromOriginal) {
-    if (!isObjectVersion(modelName)) {
-      return objectVersion;
-    }
-    originalModelName = modelName.split('_')[0];
 
-    return {
-      originalModelName: originalModelName,
-      widgetId: modelName,
-      widgetName: businessModels[originalModelName].title_plural +
-        ' Versions',
-    };
+  if (!isObjectVersion(modelName)) {
+    return objectVersion;
   }
+  originalModelName = modelName.split('_')[0];
 
   return {
-    originalModelName: modelName,
-    widgetId: modelName + '_version',
-    widgetName: businessModels[modelName].title_plural +
+    originalModelName: originalModelName,
+    isObjectVersion: !!originalModelName,
+    widgetId: modelName,
+    widgetName: businessModels[originalModelName].title_plural +
       ' Versions',
   };
-}
-
-function getWidgetConfig(modelName, buildVersionFromOriginal) {
-  let config;
-  let isObjectVersion;
-  let originalModelName;
-  let configObject;
-
-  // Workflow approach
-  if (_.isObject(modelName)) {
-    modelName.widgetName = modelName.name;
-    modelName.widgetId = modelName.name;
-    return modelName;
-  }
-
-  if (cachedObjects[modelName]) {
-    return cachedObjects[modelName];
-  }
-
-  config = _getObjectVersionConfig(modelName, buildVersionFromOriginal);
-  isObjectVersion = !!config.originalModelName;
-  originalModelName = config.originalModelName || modelName;
-
-  configObject = {
-    name: originalModelName,
-    widgetId: config.widgetId || modelName,
-    widgetName: config.widgetName || modelName,
-    countsName: modelName,
-    isObjectVersion: isObjectVersion,
-  };
-
-  cachedObjects.modelName = configObject;
-  return configObject;
-}
-
-function getWidgetConfigs(modelNames) {
-  let configs = modelNames.map(function (modelName) {
-    return getWidgetConfig(modelName);
-  });
-  return configs;
 }
 
 export {
   isObjectVersion,
   parentHasObjectVersions,
-  getWidgetConfig,
-  getWidgetConfigs,
+  getObjectVersionConfig,
 };
